Address deadlock when fetching pages.
The default size for a channel is 0, so anything sending on a
default-sized channel will block until something else reads from the
other side. Since workers read from the Jobs channel, but also send on
the channel in the event of a failure, if all of them end up failing on
a page and attempting to send the failed page through the same channel,
there'll be no workers left to read from the channel and the whole
process will hang. Increase the channel size to the expected number of
pages to reduce the chance of workers blocking themselves.